requirements.txt
lxml=4.5.0
python-docx=0.8.10
requests=2.23.0
